getting unum, array, and csr_matrix working with the schema library
