+++ genEcoli/__init__.py
@@ -9,8 +9,9 @@ from ecoli.library.schema import (
     bulk_numpy_updater,
 )
 
-from genEcoli.infer import infer_representation, infer_schema, ECOLI_TYPES, MISSING_TYPES
+from genEcoli.infer_representation import infer_representation, infer_schema, ECOLI_TYPES, MISSING_TYPES
 from genEcoli.interface import OmniStep, OmniProcess, update_inheritance, scan_processes, update_processes, migrate_composite
+import genEcoli.types 
 
 
 def register_types(core):
