+++ genEcoli/types/__init__.py
@@ -1,0 +1,2 @@
+from genEcoli.types.unum import UnumUnits
+from genEcoli.types.csr_matrix import CSRMatrix
